Exhaustive single-threaded search is done

+++ CAESAR/src/operator/ContextInitiation.java
@@ -18,7 +18,7 @@ public class ContextInitiation extends Operator {
 		return this.equals(neighbor);
 	}
 
-	public int getCost() {
+	public double getCost() {
 		return 1;
 	}
