+++ CAESAR/src/operator/RunUpdate.java
@@ -26,7 +26,7 @@ public class RunUpdate extends Operator {
 		return this.equals(neighbor);		
 	}
 	
-	public int getCost() {
+	public double getCost() {
 		return tuples.size();
 	}
